requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
langchain
langchain
langchain-core
langchain-community
langchain-groq